Update nav layout and styles. Cleanup unused css. (#71)
The goal of this PR is to update the main navigation's layout and styles in accordance with the mockups created by @edamamedesign. The orgLogoUrl still needs to be tweaked so uploaded photos display nicely. For now, only the `username` is displayed at the top of the navigation and not the `orgName`.