Updated .dockerignore and added non-root user to Dockerfile (#143)
Running applications as root (default Docker user) is not a good security practice. The goal of this PR is to run the Fleet application as a non-root user. 

For development you can still access the container as root with the following command: `docker run -it --user=0 <fleetdm container ID> bash`