Update fleetctl NPM package to not use postinstall
We found in some environments users run `sudo npm install -g fleetctl`
to install via NPM, and NPM does not execute postinstall scripts when
run as root.

This new approach should be compatible with more environments as the
download of fleetctl is done as necessary when the script is invoked.