Delete package-lock.json
Ack! I committed package-lock.json on my first commit too. It's naturally the right thing to do but the team at fleet agreed they don't want to commit package-lock.json for the website. Zach's going to go ahead and add it to the .gitignore file so it won't happen again.