Remove usage and packaging documentation. Link to Fleet repo (#22)
The goal of this PR is to move the "Usage" and "Packaging" sections of the Orbit documentation to the Fleet repo. This way, Fleet users can remain in the Fleet docs when reading Orbit documentation. Eventually, when the Fleet docs move to fleetdm.com, the Orbit documentation will live on the website as well.