Handbook style fix
Reduced handbook template width to bring inline with the width in Figma. (The extra 200px forces the text content a little too wide, introducing reading fatigue.)

Reverted paragraph font size back to default 16px, and also removed the bootstrap override to avoid duplicate declarations.

Thank you @eashaw for humoring me with the font-size change. I realise now that it's not the size, but the actual font that is the issue. Looks like there is a discrepancy between how Figma and the browser are rendering Nunito (specifically the curly tails on the 'l'). I will track this separately for later.