Handbook: Blog post publishing process (#2299)
* Update growth.md

Added a section for Fleet blog post authors about how to promote their blog posts to our social media.

Some thoughts. Including an assignee is not very future proof, in case roles change etc. Maybe best to leave that out, and we can add an assignee when we see the issue in Growth meetings.

HackerNews reserved for release posts only. Added that based on recent recommendation, due to HackerNew's strict policies. Are we happy with that, or should promoting on HackerNews be treated case-by-case?

* Update growth.md

* Update growth.md

* Update growth.md

* Update growth.md