Handbook Ritual Workiversaries (#4978)
* Handbook Ritual Workversaries 

Moved Slack channels heading down and added a ritual heading for "Workversaries."
Added a description line. 
Added two lines summarizing the ritual.

* Update people.md

* Revised edits

Line 270 Changed Workversaries to Workiversaries
Line 274 traded out the list for a sentence summary of the Ritual. 

Should reference to Mike be they/them or he/him in this context?

Co-authored-by: Mike McNeil <mikermcneil@users.noreply.github.com>