Add initial community ritual table to the handbook (#5159)
* feat: add initial community ritual table to the handbook

* fix: retype markdown pipes and dashes

* fix: remove extra rows and columns

* fix: remove tweet ritual

* Update community.md

I made one change and moved the table up. The Slack section should always be at the bottom.

* Update community.md

I made one change. Remember, when a comma or period goes at the end of a quote it should always be inside of the quotations. This does not apply to exclamation points or question marks for the sentence outside of the quotes.

Co-authored-by: Desmi-Dizney <99777687+Desmi-Dizney@users.noreply.github.com>