Editor pass for capitalization
Headings should only have the first word capitalized unless something is a proper noun or represents an acronym. The Rituals section is already on the Growth page, so out of consistency across all Handbook pages, it needs only to be given the heading "Rituals." This helps keep URLs consistent as well.