Update customers.md (#6076)
I'd like to spend some time to make the customer page more concise. First steps have been actioned in this PR as follows:

• Updated the page hierarchy with H2 sections
• Reformatted the "Customer success call" section.

I would like to tackle the rest of the page and have scheduled time with @tgauda to understand the concepts described on the page in order to lay it out better.