Website: Follow up to strip leading slashes (#6843)
* Website: Follow up to strip leading slashes

re https://github.com/fleetdm/fleet/pull/6796#issuecomment-1193054810

* remove log

* Verified: whitespace trim not needed

And leading slash trim is only needed if using a regex route (not something I recommend most of the time).  But since we are using one, I included it, and put it in all three places for consistency.  (It doesn't ever hurt)