Article: Osquery as a threat hunting platform (#7626)
* Article: Osquery as a threat hunting platform

* Add .md to title

We need to define the file type in the title. These articles are structured in markdown, so we add ".md" to the title.

* Add metadata

Reminder to change publishing date and article image once we have it.

* Making copy less formal

Simplified language and provided more context for osquery being a lightweight agent.

* Add images and meta tags

I made the following updates to this article:
- Added images.
- Updated category to "Security".

* Osquery threat hunting article revisions.md

Revised copy for simplicity and flow. Added a link to the first podcast episode. Followed @mike-j-thomas's suggestion to mention value of osquery managers.

Co-authored-by: Andrew Baker <89049099+DrewBakerfdm@users.noreply.github.com>
Co-authored-by: Mike Thomas <mthomas@fleetdm.com>