fix issue in fleet desktop causing it to spam free installations
the `switch` statement that checks for errors (including license errors)
issues a `continue` before we even have the chance to wait for the
ticker.

this has the drawback that premium users will have to wait 5 minutes
before they see policy info, but the alternative would be to use labels
and go-to, at least with the current code structure.