update navigation
## The problem

After consolidating Growth, Community, and Digital Experience (Brand) under "Marketing", the side navigation has become overwhelming and unclear what sections falls into each category.

## How to fix

Since the navigation can only display top-level sections, I propose splitting the navigation as follows:

- Marketing
- Community
- Brand
- Rituals
- Slack channels

Navigating to each of these sections will then display a content table for each, labeled "In this section" (as currently exists).

To do this, I made the following changes:

- Changed heading hierarchy throughout page
- added styling to h3 to include border-top

## Caveat

Since h2 is used to define the three categories on this page (Marketing, Community, Brand), h3 tag is now utilized to define sub-sections for each of the main categories. However, this styling is specific to the Marketing page of the handbook, and should not be applied to the other pages – which I do not know how to achieve.