fix integration.yml CI workflow by setting a Go version
We don't have a `matrix.go-version` defined, so the task was using
`go1.17.13` as the default.

This explicitly sets the version to `go1.19.1`, I didn't use a `matrix`
definition because at least for now, we only want to run this test using a
single Go version.

Since I was there, I also updated test-native-tooling-packaging.yml to
use `go1.19.1` too.