"Why open source?" (#9233)
In the footer, instead of https://fleetdm.com/transparency, include a
link to ["Why open
source?"](https://fleetdm.com/handbook/company/why-this-way#why-open-source)
(in "Why this way?", in the handbook)

Why? "Transparency" is a page designed especially for end users, and
linked to already from within Fleet Desktop. Let's use this footer link
to help communicate some of the same ideas, but with additional context
for IT and security folks within customer and contributor organizations.