Add macOS CIS 2.3.3.11 (Bluetooth Sharing) (#9688)
This adds a test for 2.3.3.11, which verifies that all users have
Bluetooth sharing disabled.

It's not possible to create a profile and it's challenging to write a
script, so testing must be done manually.
As with all sharing tests, the machine will be compliant out of the box.
For QA you should open the sharing preference pane (System Settings >
General > Sharing) and then enable Bluetooth sharing and this test
should fail. Switching bluetooth sharing back off will fix it.