feat(terraform): add zone_id variable to ses module
The zone_id variable is added to the ses module to allow the module to be used with different Route53 zones. The variable is used in the aws_route53_record resource to set the zone_id attribute. The aws_route53_zone data source is removed from the module and the zone_id attribute is set directly. The count attribute is added to the aws_route53_record resource to allow for multiple DKIM records to be created.