fix flaky TestIntegrationsMDM/TestDEPProfileAssignment
This test was failing because we can send other commands when a DEP
device is enrolling (eg: to install a bootstrap package) and the test
was assuming that the first command we'll always get is the command to
install `fleetd`.

We have discussed doing this in the past (@gillespi314 suggested that is
important) increasing the accuracy of the timestamps in the
`nano_commadns` table, which is the proper fix for this.